1. 修复CSSbug，更新Webfolder Grade字段 2. 先提交一次，以免丢失数据 3. 准备做预览HTML页面和评论 4. 希望今天能完成预览以及评论

+++ WebBookmarkSolution/WebBookmarkBo/Model/BizUserRelationship.cs
@@ -31,7 +31,7 @@ namespace WebfolderBo.Model
         /// <summary>
         /// 是否互相关注
         /// </summary>
-        public sbyte IsMutuallyFollwe { get; set; }
+        public int IsMutuallyFollwe { get; set; }
 
         /// <summary>
         /// 创建时间
