+++ WebBookmarkSolution/WebBookmarkService/DAL/UserRelationshipDAL.cs
@@ -121,7 +121,7 @@ namespace WebBookmarkService.DAL
 			userRelationship.UserRelationshipID = (long)ToModelValue(dr,"UserRelationshipID");
 			userRelationship.FollowerID = (long)ToModelValue(dr,"FollowerID");
 			userRelationship.BeFollwedUID = (long)ToModelValue(dr,"BeFollwedUID");
-			userRelationship.IsMutuallyFollwe = (sbyte)ToModelValue(dr,"IsMutuallyFollwe");
+			userRelationship.IsMutuallyFollwe = (int)ToModelValue(dr,"IsMutuallyFollwe");
 			userRelationship.CreateTime = (DateTime)ToModelValue(dr,"CreateTime");
 			return userRelationship;
 		}
